SNEK-248_user-authentication/django-react-session-cookies/server/requirements.txt
Django==4.0.1
djangorestframework==3.13.1
django-cors-headers==3.11.0